requirements.txt
google-generativeai
google-adk
pandas
openpyxl
requests
beautifulsoup4
db-sqlite3